requirements.txt
requests
beautifulsoup4
python-dotenv
schedule
spotipy
numpy
Pillow
scikit-learn
io
